Refactor configuration variable lookup
Change the `Application.get_env/3` calls to use a new function
`Recaptcha.Config.get_env/3` to perform the lookups. Doing so allows the
project to defer environment variable lookup to runtime instead of only at
compile-time.

For example, we can have the configuration be specified as each of the
following:

*   Standard variables:

    config :recaptcha,
        public_key: "some public key string",
        secret: "some private string"

*   Environment variable at compile time:

    config :recaptcha,
        public_key: System.get_env("RECAPTCHA_PUBLIC_KEY"),
        secret: System.get_env("RECAPTCHA_PRIVATE_KEY")

*   Environment variable at runtime:

    config :recaptcha,
        public_key: {:system, "RECPATCHA_PUBLIC_KEY"},
        secret: {:system, "RECAPTCHA_PRIVATE_KEY"}